requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
python-decouple==3.8
openai==1.3.7
asyncio
aiohttp==3.9.1
requests==2.31.0
deep-researcher==0.0.10
django-allauth==0.57.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1 